Added TODO parsing from a Git repo
When passed the -g flag, twinput will now read comments in a Git repo
using the git grep command. The comments will then be formatted
according to a template file and added to the Taskwarrior database.

The syntax for adding a task has also been modified to be more
compliant with the Todo.txt standard. Quotation marks around the task
description are no longer required.

twinput will also generate an MD5 hash of the task description when
adding a task. The hash is stored in the Taskwarrior database as a UDA.
If twinput can retrieve a task with a matching hash, then twinput will
just update the task rather than creating a new one. This is primarily
intended to keep the gitgrep mode from parsing the same comment into
the database several times.

When in interactive mode, twinput will collect any errors generated
while parsing and re-open the input buffer presenting the errors and
their error messages to the user for remodification. As a part of this
more elegant error handling process, errored tasks are no longer
deleted when parsing fails, but rather the hashing functionality above
allows the user to update or overwrite.