Fixed no-open bug
The text editor was not opening at all when no arguments were being
passed. In this fix, no arguments now means the text editor opens while
one or more arguments will either change the behavior of the script or
throw a usage error.

There was also a debug print statement showing the output of a
git-grep'd template file. This debug output was removed.