Did some needed refactoring
Since TWInput is starting to grow up to be a full-featured app, I'm
starting the process of treating it as such. So, in this commit, I
factored everything out into a couple of modules. More refactoring
might be called for in the future, but this is a start. No features
were actually added here.