Fix tasks not updating
Had the loop that updated the task dictionary hidden under the wrong conditional.

This could still be cleaner. This will perform a few operations per task which will clutter the TaskWarrior database, but the way TaskWarrior is set up, I'm starting to think this is 1) unavoidable, and 2) not really that big of a deal.