Added task review tool
Reworked the interactive (-i) feature, which up to this point was
compeletly broken and, therefore, pointless, to act as a task
review tool. It will iterate over every task in the database and
give the user the option to edit effort estimates and toggle
between the someday lists.

Also cleaned up command line parsing with docopt.